refactor(services): extract web code index client
Sixteen retrieval and code index methods move into `web-code-index-client.ts`,
with their four `let` bindings and the six helpers around them in
`web-code-index-state.ts`, unexported and reached through accessors.

`createSession` stays in the composition root and still needs automatic index
discovery, which reads the retrieval configuration and mints a workspace id. It
gets `discoverWebSessionCodeIndex` — the behavior, not the bindings — so the cut
holds without exporting anything mutable.

`resetWebRetrievalForTest` and `searchWebCodeIndex` are re-exported from
`web-agent-client.ts` so the existing test seams keep their import path.
